requirements.txt
python-telegram-bot==20.6
opencv-python>=4.8.0
numpy>=1.26.0
Pillow>=10.0.0
pdf2image>=1.17.0
svgwrite>=1.4.3
cairosvg>=2.7.1
reportlab>=4.1.0
python-dotenv>=1.0.1
telegram
cairosvg
numpy
opencv-python
pdf2image
pillow
pypdf2
python-dotenv
python-telegram-bot
svgwrite
